requirements.txt
fastapi
pydantic[dotenv]
uvicorn
python-dotenv
langgraph
langchain_community
duckduckgo-search
langchain-openai
langchain-tavily
langchain-groq
pytz
boto3
bson
pymongo
python-jose
passlib
pydantic[email]
python-multipart